requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.1
bleach==3.3.0
camelot-py==0.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.5
cryptography==3.1.1
distlib==0.3.0
distro==1.5.0
dj-database-url==0.5.0
Django==3.1.7
django-crontab==0.7.1
django-filter==2.4.0
django-heroku==0.3.1
django-mailer==2.1
django-qr-code==2.1.0
django-sslserver==0.22
django-stubs==1.7.0
djangorestframework==3.12.2
docutils==0.16
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
isort==5.5.3
jdcal==1.4.1
keyring==22.3.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
mysql-connector-python==8.0.20
numpy==1.19.2
openpyxl==3.0.5
packaging==20.9
pandas==1.1.2
pdfminer.six==20200726
pkginfo==1.7.0
psycopg2==2.8.6
pycparser==2.20
Pygments==2.8.0
pylint==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
segno==1.3.1
selenium==3.141.0
six==1.14.0
sortedcontainers==2.2.2
soupsieve==2.0.1
sqlparse==0.4.1
tabula==1.0.5
tabula-py==2.2.0
toml==0.10.1
tqdm==4.58.0
twine==3.3.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==20.0.20
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
